requirements.txt
pystray==0.19.5
Pillow>=9.3
win10toast==0.9
